Split Android notification workflows
