+++ apps/android/app/src/main/java/com/flashcardsopensourceapp/app/di/AppGraph.kt
@@ -24,9 +24,9 @@ import com.flashcardsopensourceapp.core.ui.AppMessageBus
 import com.flashcardsopensourceapp.core.ui.TestModeStore
 import com.flashcardsopensourceapp.core.ui.VisibleAppScreenController
 import com.flashcardsopensourceapp.app.navigation.AppHandoffCoordinator
-import com.flashcardsopensourceapp.app.notifications.ReviewNotificationsManager
-import com.flashcardsopensourceapp.app.notifications.AndroidStrictRemindersScheduler
-import com.flashcardsopensourceapp.app.notifications.StrictRemindersManager
+import com.flashcardsopensourceapp.app.notifications.review.ReviewNotificationsManager
+import com.flashcardsopensourceapp.app.notifications.strict.AndroidStrictRemindersScheduler
+import com.flashcardsopensourceapp.app.notifications.strict.StrictRemindersManager
 import com.flashcardsopensourceapp.data.local.bootstrap.ensureLocalWorkspaceShell
 import com.flashcardsopensourceapp.data.local.ai.remote.AiChatLiveRemoteService
 import com.flashcardsopensourceapp.data.local.ai.store.AiChatHistoryStore
