+++ apps/android/app/src/main/java/com/flashcardsopensourceapp/app/notifications/NotificationPermission.kt
@@ -1,0 +1,13 @@
+package com.flashcardsopensourceapp.app.notifications
+
+import android.Manifest
+import android.content.Context
+import android.content.pm.PackageManager
+import androidx.core.content.ContextCompat
+
+fun hasNotificationPermission(context: Context): Boolean {
+    return ContextCompat.checkSelfPermission(
+        context,
+        Manifest.permission.POST_NOTIFICATIONS
+    ) == PackageManager.PERMISSION_GRANTED
+}
