+++ apps/android/app/src/main/java/com/flashcardsopensourceapp/app/notifications/ReviewNotificationWorker.kt
@@ -1,93 +1,13 @@
 package com.flashcardsopensourceapp.app.notifications
 
 import android.content.Context
-import androidx.work.CoroutineWorker
 import androidx.work.WorkerParameters
-import com.flashcardsopensourceapp.app.FlashcardsApplication
-import com.flashcardsopensourceapp.data.local.notifications.ReviewNotificationsStore
-import com.flashcardsopensourceapp.data.local.notifications.SharedPreferencesReviewNotificationsStore
+import com.flashcardsopensourceapp.app.notifications.review.ReviewNotificationWorker as ReviewNotificationWorkflowWorker
 
+/**
+ * Compatibility entry point for WorkManager rows persisted before the notification package split.
+ */
 class ReviewNotificationWorker(
     appContext: Context,
     params: WorkerParameters
-) : CoroutineWorker(appContext, params) {
-    override suspend fun doWork(): Result {
-        val requestId = inputData.getString(reviewNotificationRequestIdDataKey)
-        val workspaceId = inputData.getString(reviewNotificationWorkspaceIdDataKey)
-        val permissionAllowed: Boolean = hasNotificationPermission(context = applicationContext)
-        addWorkerBreadcrumb(
-            stage = "worker_start",
-            requestId = requestId,
-            workspaceId = workspaceId,
-            permissionAllowed = permissionAllowed
-        )
-        if (permissionAllowed.not()) {
-            addWorkerBreadcrumb(
-                stage = "worker_permission_blocked",
-                requestId = requestId,
-                workspaceId = workspaceId,
-                permissionAllowed = permissionAllowed
-            )
-            return Result.success()
-        }
-
-        val frontText = inputData.getString(reviewNotificationFrontTextDataKey)
-        if (frontText == null || requestId == null || workspaceId == null) {
-            addWorkerBreadcrumb(
-                stage = "worker_invalid_input_failure",
-                requestId = requestId,
-                workspaceId = workspaceId,
-                permissionAllowed = permissionAllowed
-            )
-            return Result.failure()
-        }
-
-        // Read live so a toggle-off after schedule time wins immediately, even if a
-        // worker is already mid-flight.
-        val store = resolveReviewNotificationsStore()
-        val showAppIconBadge = store.loadSettings(workspaceId = workspaceId).showAppIconBadge
-
-        showReviewReminderNotification(
-            context = applicationContext,
-            frontText = frontText,
-            requestId = requestId,
-            showAppIconBadge = showAppIconBadge
-        )
-        addWorkerBreadcrumb(
-            stage = "worker_notification_posted",
-            requestId = requestId,
-            workspaceId = workspaceId,
-            permissionAllowed = permissionAllowed
-        )
-        return Result.success()
-    }
-
-    private fun resolveReviewNotificationsStore(): ReviewNotificationsStore {
-        val appGraphStore = (applicationContext as? FlashcardsApplication)
-            ?.appGraphOrNull
-            ?.reviewNotificationsStore
-        if (appGraphStore != null) {
-            return appGraphStore
-        }
-        // Cold-start fallback: the worker can fire before Application.onCreate has published the graph.
-        return SharedPreferencesReviewNotificationsStore(context = applicationContext)
-    }
-
-    private fun addWorkerBreadcrumb(
-        stage: String,
-        requestId: String?,
-        workspaceId: String?,
-        permissionAllowed: Boolean
-    ) {
-        addNotificationWorkerBreadcrumb(
-            applicationContext = applicationContext,
-            notificationKind = reviewReminderNotificationKind,
-            stage = stage,
-            requestId = requestId,
-            workspaceId = workspaceId,
-            permissionAllowed = permissionAllowed,
-            workTag = reviewNotificationWorkTag,
-            workLimit = null
-        )
-    }
-}
+) : ReviewNotificationWorkflowWorker(appContext, params)
