+++ apps/android/app/src/test/java/com/flashcardsopensourceapp/app/notifications/ReviewNotificationsManagerTest.kt
@@ -1,5 +1,10 @@
 package com.flashcardsopensourceapp.app.notifications
 
+import com.flashcardsopensourceapp.app.notifications.review.ReviewNotificationFilterPlan
+import com.flashcardsopensourceapp.app.notifications.review.resolveExactStoredReviewTagNames
+import com.flashcardsopensourceapp.app.notifications.review.resolveReviewNotificationFilterPlan
+import com.flashcardsopensourceapp.app.notifications.review.reviewReminderNotificationTag
+import com.flashcardsopensourceapp.app.notifications.strict.strictReminderNotificationTag
 import com.flashcardsopensourceapp.data.local.model.cards.DeckFilterDefinition
 import com.flashcardsopensourceapp.data.local.model.review.ReviewFilter
 import org.junit.Assert.assertEquals
